added_trace_ratio
