+++ ITMO_FS/filters/__init__.py
@@ -1,2 +1,3 @@
 from .multivariate import *
 from .univariate import *
+from .unsupervised import *
