+++ ITMO_FS/filters/multivariate/__init__.py
@@ -2,4 +2,4 @@ from .DISRwithMassive import *
 from .FCBF import *
 from .MultivariateFilter import MultivariateFilter
 from .measures import *
-from .TraceRatio import TraceRatio
+from .trace_ratio_fisher import TraceRatioFisher
