+++ ITMO_FS/filters/unsupervised/__init__.py
@@ -1,0 +1,1 @@
+from .trace_ratio_laplacian import TraceRatioLaplacian
